extract bucket and file classes to be on their own
